+++ fluent_pages/views/dispatcher.py
@@ -3,7 +3,6 @@ The view to display CMS content.
 """
 from django.conf import settings
 from django.core.urlresolvers import Resolver404, reverse, resolve, NoReverseMatch
-from django.core.xheaders import populate_xheaders
 from django.http import Http404, HttpResponseRedirect
 from django.views.generic.base import View
 from fluent_pages.models import UrlNode
@@ -128,9 +127,6 @@ class CmsPageDispatcher(GetPathMixin, View):
             # Avoid automatic fallback to 404 page in this dispatcher.
             raise ValueError("The method '{0}.get_response()' didn't return an HttpResponse object.".format(plugin.__class__.__name__))
 
-        if 'X-Object-Type' not in response:
-            populate_xheaders(self.request, response, self.object, self.object.pk)
-
         return response
 
 
@@ -184,9 +180,6 @@ class CmsPageDispatcher(GetPathMixin, View):
         response = match.func(self.request, *match.args, **match.kwargs)
         if response is None:
             raise RuntimeError("The view '{0}' didn't return an HttpResponse object.".format(match.url_name))
-
-        if 'X-Object-Type' not in response:
-            populate_xheaders(self.request, response, self.object, self.object.pk)
 
         return response